Improve legacy field setup
The previous method (iterate once at startup) didn't handle the case when additional fields are mounted at runtime.

A MutationObserver on the parent element of editor fields seemed to be the most elegant approach. It assumes a certain HTML structure, but since the method is only called on legacy Anki versions, that shouldn't be an issue.